requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic[email]>=2.10.0  # Includes email-validator for EmailStr
pydantic-settings>=2.1.0
python-multipart>=0.0.6  # For file uploads

# Anthropic API
anthropic>=0.18.1

# Template Engine
jinja2>=3.1.6

# Environment Variables
python-dotenv==1.0.0

# HTTP Client
httpx==0.26.0

# ASGI Server
gunicorn>=22.0.0

# Google Cloud
google-cloud-secret-manager==2.18.0
google-cloud-firestore==2.14.0  # Optional: if using Firestore
google-cloud-storage>=2.14.0  # For file uploads (GCS)
google-cloud-tasks>=2.16.0  # For background processing

# Google Authentication (for IAP)
google-auth>=2.27.0
google-auth-httplib2>=0.2.0

# Logging
python-json-logger==2.0.7

# Rate Limiting
slowapi==0.1.9
redis>=5.0.0  # For distributed rate limiting

# Email Services (Optional - choose one based on EMAIL_SERVICE_PROVIDER)
sendgrid>=6.11.0  # For SendGrid email service
boto3>=1.34.0  # For AWS SES email service

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Development
black>=24.3.0
flake8==7.0.0
mypy==1.8.0

# Requests (HTTP library - may be used by dependencies)
# Note: Docker health check now uses curl instead of requests
requests>=2.32.4

# Text Extraction
PyMuPDF>=1.23.0  # PDF text extraction
python-docx>=1.1.0  # DOCX text extraction
pytesseract>=0.3.10  # OCR for images
Pillow>=10.0.0  # Image processing for OCR
beautifulsoup4>=4.12.0  # HTML parsing
